backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.0
youtube-transcript-api==0.6.2
PyPDF2==3.0.1
langchain==0.3.0
langchain-openai==0.2.0
supabase==2.9.1
pgvector==0.3.5
psycopg2-binary==2.9.10
python-multipart==0.0.12
pydantic==2.9.0
tiktoken==0.8.0
